chore: Move containerized tests to its own test collection
